Enemy patrolling. Spline patrol actor. Random wait. Immobilising enemy with chain.
Enemy patrolling behaviour tree.
Enemy patrol spline actor.
Enemy patrol component that handles the destination calculation for the patrolling and also where to start on the spline patrol depending on starting location.
3 tasks: random wait, update patrol loc and find starting patrol point which are all self-explanatory.

Immobilise component that has 2 functions: immobilise and stop immobilise. Immobilisation works by disabling the movement which is setting the movement mode to none so then stopping the immobilising is done by setting the movement mode back to walking. This works the same for the random wait task.